Load widget specific css on quickedit form.
We were defining both javascript and css of the addable widgets,
but were only adding the javascript, not the css.

The calendar widget is still not really working there.  But if we are
loading this css when viewing a form, we should load it while
quickediting the form.  On Plone 4.3 it helped.