In tests we require plone.protect>=2.0.2.
This has the 'createToken' function we use.
This is included since Plone 4.2.3.

Should be fine to use on earlier Plones, including 4.1, which uses plone.protect 2.0.
plone.protect 2.0 is fine in normal use, we just need the helper function in the tests.

In Travis we now explicitly set the plone.protect version to 2.0.