Fix redundant template fallback value causing errors
`|nothing` in conjunction with `structure` or `provider:` (not sure which)
caused `ContentProviderLookupError: archetypes.edit.beforefieldsets|nothing`
for me in some cases while debugging broken `RichTextField`.
It seems that it might not be a valid syntax?

In any case, it seems to be redundant, since there is also
`tal:on-error="nothing"`, and subsequent equicalent
`tal:replace="provider:archetypes.edit.afterfieldsets"`
does not have it either.