Fixed portal_skins directory path on Plone 5.2
It was `PloneFormGen/skins/PloneFormGen` but now has to be `Products.PloneFormGen:skins/PloneFormGen`.
This might need an upgrade step, but you can also do this manually in portal_skins in the ZMI.
Note that PloneFormGen is not really tested on Plone 5.2.
In fact the automatic tests completely fail on 5.2, so use at your own risk.